bq_to_sendgrid/requirements.txt
google-cloud-bigquery
sendgrid
